:art: feat: 단위 변환 구현

+++ backend/src/main/java/com/donjo/backend/solidity/support/SupportSol.java
@@ -19,7 +19,7 @@ public class SupportSol {
 
     public String to;
 
-    public Double amount; // matic
+    public Long amount; // matic
 
     public LocalDateTime sendTimestamp;
